The `Call<T, S>` struct has been removed. It was the basis for the routing mechanism based on `TypeId`s.
This mechanism limits the number of arguments of handler functions to
one argument: some instance of `Call<T, S>`. For example,
`Call<Authorize, CGW>`.

Registering a handler function with more arguments would probably
compile. However, the route would never be called.

In the old logic, a route is identified by the `TypeId` 'calculated'
from all the arguments of the handler functions. The `Router` stores the
handler function in a `HashMap` using this `TypeId` as a key.

However, when a request comes in and the `Router` creates an instance of
`Call<T, S>` and 'calculates' it's `TypeId`. Next, `Router` looks up the
`TypeId` in the `HashMap` and executes the handler functions.

If the handler function has more arguments than `Call<T, S>`, it's never
retrieved from the `HashMap`.

Because `Call<T, S>` is removed, I rewrote `Request` to a regular
struct.